backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0 
langchain==0.3.26
langchain-community==0.3.27
langchain-google-genai>=2.1.6
google-genai==1.24.0
pgvector==0.2.4
psycopg[binary]==3.1.18
python-dotenv==1.0.1
pypdf==4.2.0
python-multipart==0.0.6
slowapi==0.1.9
secure==1.0.1
